# Conflicts:
#	sdk/iOS/OoyalaSkinSDK.xcodeproj/project.pbxproj
#	sdk/iOS/OoyalaSkinSDK/React/OOReactSkinModel.h
#	sdk/iOS/OoyalaSkinSDK/React/OOReactSkinModel.m
#	sdk/iOS/OoyalaSkinSDK/ViewControllers/OOSkinViewController.m
#	sdk/iOS/OoyalaSkinSDK/Views/CountdownView.h
#	sdk/iOS/OoyalaSkinSDK/Views/CountdownView.m
#	sdk/iOS/OoyalaSkinSDK/Views/CountdownViewManager.m
#	sdk/iOS/OoyalaSkinSDK/Views/OOClosedCaptionsViewManager.h
#	sdk/iOS/OoyalaSkinSDK/Views/OOClosedCaptionsViewManager.m
#	sdk/iOS/OoyalaSkinSDK/Views/OOGradientView.h
#	sdk/iOS/OoyalaSkinSDK/Views/OOGradientView.m
#	sdk/iOS/OoyalaSkinSDK/Views/OOGradientViewManager.h
#	sdk/iOS/OoyalaSkinSDK/Views/OOGradientViewManager.m
#	sdk/iOS/OoyalaSkinSDK/Views/OOVolumeViewManager.h
#	sdk/react/bottomOverlay.js
#	sdk/react/common/progressBar.js
#	sdk/react/constants.js
#	sdk/react/ooyalaSkinBridgeListener.js
#	sdk/react/ooyalaSkinPanelRenderer.js
#	sdk/react/panels/AudioAndCCSelectionPanel.js
#	sdk/react/panels/videoView.js
#	sdk/react/widgets/controlBarWidgets.js

+++ sdk/iOS/OoyalaSkinSDK/OOSkinPlayerObserver.m
@@ -2,7 +2,7 @@
 //  OOSkinPlayerObserver.m
 //  OoyalaSkinSDK
 //
-//  Created by Michael Len on 2/25/16.
+//  Created on 2/25/16.
 //  Copyright © 2016 ooyala. All rights reserved.
 //
 
@@ -358,7 +358,7 @@ static NSString *requireAdBarKey = @"requireAdBar";
 
   NSInteger count       = [adInfo[countKey] integerValue];
   NSInteger unplayed    = [adInfo[unplayedKey] integerValue];
-  NSString *countString = [NSString stringWithFormat:@"(%ld/%ld)", count - unplayed, (long)count];
+  NSString *countString = [NSString stringWithFormat:@"(%ld/%ld)", (long)(count - unplayed), (long)count];
   NSNumber *skipoffset  = @([adInfo[skipOffsetKey] floatValue]);
   NSArray *icons        = adInfo[iconsKey];
   NSString *title       = adInfo[titleKey];
